+++ matrix_functions.h
@@ -5,8 +5,8 @@
 #ifndef MATRIX_FUNCTIONS_H
 #define MATRIX_FUNCTIONS_H
 
-float* transpose(const float* inmatrix, int rowcount, int columncount) {
-    float* out = (float*)malloc(sizeof(float) * rowcount * columncount);
+double* transpose(const double* inmatrix, int rowcount, int columncount) {
+    double* out = (double*)malloc(sizeof(double) * rowcount * columncount);
 
     for (int row = 0; row < rowcount; row++) {
         for (int col = 0; col < columncount; col++) {
@@ -17,21 +17,21 @@ float* transpose(const float* inmatrix, int rowcount, int columncount) {
     return out;
 }
 
-void* scalar_multiply(void* matrix, int rowcount, int columncount, float scalar) {
-    void* outmatrix = calloc(rowcount * columncount, sizeof(float));
+void* scalar_multiply(void* matrix, int rowcount, int columncount, double scalar) {
+    void* outmatrix = calloc(rowcount * columncount, sizeof(double));
     for (int i = 0; i < rowcount; i++) {
         for (int j = 0; j < columncount; j++) {
-            void* source = (char*)matrix + (i * columncount + j) * sizeof(float);
-            void* dest = (char*)outmatrix + (i * columncount + j) * sizeof(float);
-            *(float*)dest = (*(float*)source) * scalar;
+            void* source = (char*)matrix + (i * columncount + j) * sizeof(double);
+            void* dest = (char*)outmatrix + (i * columncount + j) * sizeof(double);
+            *(double*)dest = (*(double*)source) * scalar;
         }
     }
     return outmatrix;
 }
 
-float* hadamard_multiply(float* in1, float* in2, int rowcount, int columncount) {
+double* hadamard_multiply(double* in1, double* in2, int rowcount, int columncount) {
     int size = rowcount * columncount;
-    float* outmatrix = calloc(size, sizeof(float));
+    double* outmatrix = calloc(size, sizeof(double));
     if (!outmatrix) return NULL;
 
     for (int i = 0; i < size; i++) {
@@ -42,11 +42,11 @@ float* hadamard_multiply(float* in1, float* in2, int rowcount, int columncount) 
 }
 
 
-void* matrix_multiply(float* in1, float* in2, int row, int common, int col) {
-    float* outmatrix = calloc(row * col, sizeof(float));
+void* matrix_multiply(double* in1, double* in2, int row, int common, int col) {
+    double* outmatrix = calloc(row * col, sizeof(double));
     for (int i = 0; i < row; i++) {
         for (int j = 0; j < col; j++) {
-            float sum = 0;
+            double sum = 0;
             for (int k = 0; k < common; k++) {
                 sum += in1[i * common + k] * in2[k * col + j];
             }
